feat: 맛집 등록시 리뷰 함께 등록 처리 가능

+++ src/main/java/site/matzip/review/dto/ReviewCreationDTO.java
@@ -7,6 +7,6 @@ import site.matzip.member.domain.Member;
 public class ReviewCreationDTO {
     private Long reviewId;
     private Member author;
-    private Long rating;
+    private double rating;
     private String content;
 }
